[refactor] #151 - 게시물 API를 Post 네트워크 계층으로 통합
