+++ POTI-iOS/POTI-iOS/Data/Network/Post/DTO/PotDetailDTO.swift
