+++ POTI-iOS/POTI-iOS/Data/Network/Post/DTO/PotListDTO.swift
